+++ src/shared/ui/card/CardContent.tsx
@@ -1,0 +1,8 @@
+import { forwardRef } from "react"
+import * as React from "react"
+
+export const CardContent = forwardRef(({ className, ...props }, ref) => (
+  <div ref={ref} className={`p-6 pt-0 ${className}`} {...props} />
+))
+
+CardContent.displayName = "CardContent"
