+++ src/shared/ui/card/index.ts
@@ -1,0 +1,4 @@
+export { Card } from "./Card"
+export { CardContent } from "./CardContent"
+export { CardHeader } from "./CardHeader"
+export { CardTitle } from "./CardTitle"
