+++ src/shared/ui/dialog/index.ts
@@ -1,0 +1,9 @@
+export {
+  Dialog,
+  DialogContent,
+  DialogHeader,
+  DialogOverlay,
+  DialogPortal,
+  DialogTitle,
+  DialogTrigger,
+} from "./Dialog.tsx"
