+++ src/shared/ui/table/index.ts
@@ -1,0 +1,1 @@
+export { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from "./Table.tsx"
